+++ R/plot_predicted_curve.R
@@ -7,12 +7,12 @@
 #' and shows all individual 
 #' sampled curves.
 #'
-#' @param model An `sr_model` object (returned by [run_mod()]) containing 
+#' @param model An `sr_model` object (returned by [run_mod]) containing 
 #'   samples from the posterior distribution of the model parameters.
 #' @param ids The participant IDs to plot; for example, `"sees_npl_128"`.
 #' @param antigen_iso  The antigen isotype to plot; for example, "HlyE_IgA" or 
 #' "HlyE_IgG".
-#' @param dataset (Optional) A [tibble::tbl_df] with observed antibody response 
+#' @param dataset (Optional) A [tibble::tbl_df] with observed antibody response
 #' data. 
 #' Must contain:
 #'   - `timeindays`